--- SRC/Script/MainParts/UIParts/MainPart.ts
@@ -1,6 +1,0 @@
-import { MainPageUI } from './MainUI';
-import '../../../CSS/MainPageUI.css'
-import 'font-awesome/css/font-awesome.css'
-window.onload = function () {
-    MainPageUI.MainPageSet();
-}
