+++ package.json
@@ -4,8 +4,9 @@
   "description": "Autopilot console",
   "main": "/SRC/Script/Comped/main.js",
   "scripts": {
-    "build": "webpack --config ./webpack.config.js",
-    "start": "npm run build && electron ./SRC/Script/Comped/main.js"
+    "rebuild": "electron-rebuild -w opencv4nodejs",
+    "startdev": "electron-webpack dev",
+    "start": "electron-webpack"
   },
   "repository": {
     "type": "git",
@@ -27,21 +28,19 @@
     "@types/node": "^12.12.35",
     "@types/react": "^16.9.34",
     "@types/react-dom": "^16.9.6",
-    "css-loader": "^1.0.1",
-    "electron": "^7.1.9",
-    "file-loader": "^6.0.0",
-    "html-webpack-plugin": "^4.2.0",
-    "style-loader": "^1.2.0",
-    "ts-loader": "^7.0.1",
+    "css-loader": "^3.5.3",
+    "electron": "^8.2.5",
+    "electron-rebuild": "^1.10.1",
+    "electron-webpack": "^2.8.2",
+    "ts-loader": "^7.0.4",
     "typescript": "^3.8.3",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "url-loader": "^4.1.0"
+    "webpack": "^4.43.0"
   },
   "dependencies": {
     "echarts": "^4.7.0",
     "font-awesome": "^4.7.0",
     "leaflet": "^1.6.0",
+    "opencv4nodejs": "^4.2.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   }
