+++ tsconfig.json
@@ -5,7 +5,5 @@
         "jsx": "react",
         "target": "ES6",
         "module": "CommonJS",
-        "rootDir": "./SRC/Script/MainParts",
-        "outDir": "./SRC/Script/Comped",
     }
 }
